+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - Upgraded PHPUnit to version 10 !3092 @\_fridtjof_
 - Upgraded to Symfony 6.4 !3129 @\_fridtjof_
 - Work around a bug in Sentry's query profiling integration !3133 @\_fridtjof_
+- Added profile pictures to seed data #738 !3172 @AntonBallmaier
 
 ## Development documentation